+++ CHANGELOG.md
@@ -1,7 +1,12 @@
+## 0.36.0
+
+#### New icons
+* Added 12 new animated `clipboard` icons: `clipboard`, `clipboard-check`, `clipboard-clock`, `clipboard-copy`, `clipboard-list`, `clipboard-minus`, `clipboard-paste`, `clipboard-pen`, `clipboard-pen-line`, `clipboard-plus`, `clipboard-type`, and `clipboard-x`.
+
 ## 0.35.0
 
 #### New icons
-* Added new animated icons: `circuit-board`, `citrus`, `fclapperboard`.
+* Added new animated icons: `circuit-board`, `citrus`, `clapperboard`.
 
 ## 0.34.1